+++ src/state/viewModel/textViewModel.ts
@@ -1,5 +1,6 @@
 import type { PayloadAction } from '@reduxjs/toolkit';
 import type { AutoplayService } from '@service/autoplay';
+import type { NotificationService } from '@service/notification';
 import type { TextService } from '@service/text';
 import type { PlotState } from '@type/state';
 import type { AppStore } from '../store';
@@ -10,12 +11,14 @@ interface TextState {
   enabled: boolean;
   announce: boolean;
   value: string;
+  message: string | null;
 }
 
 const initialState: TextState = {
   enabled: true,
   announce: true,
   value: '',
+  message: null,
 };
 
 const textSlice = createSlice({
@@ -23,6 +26,7 @@ const textSlice = createSlice({
   initialState,
   reducers: {
     update(state, action: PayloadAction<string>): void {
+      state.message = null;
       state.value = action.payload;
     },
     announceText(state, action: PayloadAction<boolean>): void {
@@ -31,20 +35,28 @@ const textSlice = createSlice({
     toggle(state, action: PayloadAction<boolean>): void {
       state.enabled = action.payload;
     },
+    notify(state, action: PayloadAction<string>): void {
+      state.message = action.payload;
+    },
     reset(): TextState {
       return initialState;
     },
   },
 });
-const { update, announceText, toggle, reset } = textSlice.actions;
+const { update, announceText, toggle, notify, reset } = textSlice.actions;
 
 export class TextViewModel extends AbstractViewModel<TextState> {
   private readonly textService: TextService;
 
-  public constructor(store: AppStore, text: TextService, autoplay: AutoplayService) {
+  public constructor(
+    store: AppStore,
+    text: TextService,
+    notification: NotificationService,
+    autoplay: AutoplayService,
+  ) {
     super(store);
     this.textService = text;
-    this.registerListeners(text, autoplay);
+    this.registerListeners(notification, autoplay);
   }
 
   public dispose(): void {
@@ -52,9 +64,13 @@ export class TextViewModel extends AbstractViewModel<TextState> {
     super.dispose();
   }
 
-  private registerListeners(text: TextService, autoplay: AutoplayService): void {
-    this.disposables.push(text.onChange((e) => {
+  private registerListeners(notification: NotificationService, autoplay: AutoplayService): void {
+    this.disposables.push(this.textService.onChange((e) => {
       this.update(e.value);
+    }));
+
+    this.disposables.push(notification.onChange((e) => {
+      this.notify(e.value);
     }));
 
     this.disposables.push(autoplay.onChange((e) => {
@@ -82,6 +98,10 @@ export class TextViewModel extends AbstractViewModel<TextState> {
   public update(text: string | PlotState): void {
     const formattedText = this.textService.format(text);
     this.store.dispatch(update(formattedText));
+  }
+
+  public notify(message: string): void {
+    this.store.dispatch(notify(message));
   }
 
   private setAriaAnnouncement(enabled: boolean): void {